+++ source/services/statefsd/src/lib.rs
@@ -23,6 +23,8 @@ mod abi_seam_os;
 mod emit_os;
 #[cfg(all(feature = "os-lite", nexus_env = "os"))]
 mod os_lite;
+#[cfg(all(feature = "os-lite", nexus_env = "os"))]
+mod quota_os;
 #[cfg(all(nexus_env = "os", feature = "os-lite"))]
 mod route_os;
 #[cfg(all(nexus_env = "os", feature = "os-lite"))]